Reset input axes to make UI functional again.
I had removed all the excess inputs from the input manager, but
didn't realize that that would break the EventSystem for the UI.
It needs Submit, Cancel, Horizontal and Vertical to be defined,
otherwise it just throws an exception every frame. This was causing
the buttons not to be clickable. Fixed.